Begun adding heightmaps and balancing animation.
The balancing animation works if you disable the commented-out lines of code related to it, but it kind've "glitches" in that it often ends up happening if you happen to stand just right on a tile that's part of a strip of "floor", thinking you're on the edge for some reason. (It kind of looks like Sonic's tripping! Lol.) Because of this it's temporarily disabled.

As for the heightmaps, they now load just fine! ...That's about it, though. Actual physics using the heightmaps (Such as slopes. :smile:) are coming soon. :D